admin/hui/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-socketio[asgi]==5.11.3
Jinja2==3.1.4